requirements.txt
fastapi~=0.115.4
psycopg
psycopg2-binary
uvicorn

python-dateutil~=2.9.0.post0
redis~=5.2.0
PyJWT~=2.9.0
SQLAlchemy~=2.0.36
pika~=1.3.2
pydantic~=2.9.2
python-dotenv~=1.0.1